Move Baker and related types to separate file
