+++ backend-rust/src/graphql_api.rs
@@ -4,6 +4,7 @@
 #![allow(unused_variables)]
 
 mod account_metrics;
+mod baker;
 mod block;
 mod block_metrics;
 mod transaction;
@@ -21,11 +22,10 @@ pub(crate) use todo_api;
 use crate::{
     address::{AccountAddress, ContractAddress, ContractIndex},
     scalar_types::{
-        AccountIndex, Amount, BakerId, BigInteger, BlockHeight, DateTime, Decimal, MetadataUrl,
-        ModuleReference, TimeSpan, TransactionHash, TransactionIndex,
+        AccountIndex, Amount, BigInteger, BlockHeight, DateTime, ModuleReference, TimeSpan,
+        TransactionHash, TransactionIndex,
     },
     transaction_event::{
-        baker::BakerPoolOpenStatus,
         delegation::{BakerDelegationTarget, DelegationTarget, PassiveDelegationTarget},
         smart_contracts::InvalidContractVersionError,
         Event,
@@ -50,7 +50,6 @@ use chrono::Duration;
 use concordium_rust_sdk::{
     base::contracts_common::schema::{VersionedModuleSchema, VersionedSchemaError},
     id::types as sdk_types,
-    types::AmountFraction,
 };
 use futures::prelude::*;
 use prometheus_client::registry::Registry;
@@ -120,6 +119,7 @@ pub struct ApiServiceConfig {
 #[derive(MergedObject, Default)]
 pub struct Query(
     BaseQuery,
+    baker::QueryBaker,
     block::QueryBlocks,
     transaction::QueryTransactions,
     account_metrics::QueryAccountMetrics,
@@ -570,29 +570,6 @@ impl BaseQuery {
         }
 
         Ok(connection)
-    }
-
-    async fn baker<'a>(&self, ctx: &Context<'a>, id: types::ID) -> ApiResult<Baker> {
-        let id = IdBaker::try_from(id)?.baker_id;
-        Baker::query_by_id(get_pool(ctx)?, id).await
-    }
-
-    async fn baker_by_baker_id<'a>(&self, ctx: &Context<'a>, id: BakerId) -> ApiResult<Baker> {
-        Baker::query_by_id(get_pool(ctx)?, id).await
-    }
-
-    async fn bakers(
-        &self,
-        #[graphql(default)] _sort: BakerSort,
-        _filter: BakerFilterInput,
-        #[graphql(desc = "Returns the first _n_ elements from the list.")] _first: Option<i32>,
-        #[graphql(desc = "Returns the elements in the list that come after the specified cursor.")]
-        _after: Option<String>,
-        #[graphql(desc = "Returns the last _n_ elements from the list.")] _last: Option<i32>,
-        #[graphql(desc = "Returns the elements in the list that come before the specified cursor.")]
-        _before: Option<String>,
-    ) -> ApiResult<connection::Connection<String, Baker>> {
-        todo_api!()
     }
 
     async fn search(&self, query: String) -> SearchResult {
@@ -1864,149 +1841,6 @@ impl AccountReleaseSchedule {
     }
 }
 
-#[repr(transparent)]
-struct IdBaker {
-    baker_id: BakerId,
-}
-impl std::str::FromStr for IdBaker {
-    type Err = ApiError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        let baker_id = value.parse()?;
-        Ok(IdBaker {
-            baker_id,
-        })
-    }
-}
-impl TryFrom<types::ID> for IdBaker {
-    type Error = ApiError;
-
-    fn try_from(value: types::ID) -> Result<Self, Self::Error> { value.0.parse() }
-}
-
-struct Baker {
-    id: BakerId,
-    staked: Amount,
-    restake_earnings: bool,
-    open_status: Option<BakerPoolOpenStatus>,
-    metadata_url: Option<MetadataUrl>,
-    transaction_commission: Option<i64>,
-    baking_commission: Option<i64>,
-    finalization_commission: Option<i64>,
-}
-impl Baker {
-    async fn query_by_id(pool: &PgPool, baker_id: BakerId) -> ApiResult<Self> {
-        sqlx::query_as!(
-            Baker,
-            r#"SELECT
-    id,
-    staked,
-    restake_earnings,
-    open_status as "open_status: BakerPoolOpenStatus",
-    metadata_url,
-    transaction_commission,
-    baking_commission,
-    finalization_commission
- FROM bakers WHERE id=$1
-"#,
-            baker_id
-        )
-        .fetch_optional(pool)
-        .await?
-        .ok_or(ApiError::NotFound)
-    }
-}
-#[Object]
-impl Baker {
-    async fn id(&self) -> types::ID { types::ID::from(self.id.to_string()) }
-
-    async fn baker_id(&self) -> BakerId { self.id }
-
-    async fn state<'a>(&'a self) -> ApiResult<BakerState<'a>> {
-        let transaction_commission = self
-            .transaction_commission
-            .map(u32::try_from)
-            .transpose()?
-            .map(|c| AmountFraction::new_unchecked(c).into());
-        let baking_commission = self
-            .baking_commission
-            .map(u32::try_from)
-            .transpose()?
-            .map(|c| AmountFraction::new_unchecked(c).into());
-        let finalization_commission = self
-            .finalization_commission
-            .map(u32::try_from)
-            .transpose()?
-            .map(|c| AmountFraction::new_unchecked(c).into());
-
-        let out = BakerState::ActiveBakerState(ActiveBakerState {
-            staked_amount:    self.staked,
-            restake_earnings: self.restake_earnings,
-            pool:             BakerPool {
-                open_status:      self.open_status,
-                commission_rates: CommissionRates {
-                    transaction_commission,
-                    baking_commission,
-                    finalization_commission,
-                },
-                metadata_url:     self.metadata_url.as_deref(),
-            },
-            pending_change:   None, // This is not used starting from P7.
-        });
-        Ok(out)
-    }
-
-    async fn account<'a>(&self, ctx: &Context<'a>) -> ApiResult<Account> {
-        Account::query_by_index(get_pool(ctx)?, self.id).await?.ok_or(ApiError::NotFound)
-    }
-
-    // transactions("Returns the first _n_ elements from the list." first: Int
-    // "Returns the elements in the list that come after the specified cursor."
-    // after: String "Returns the last _n_ elements from the list." last: Int
-    // "Returns the elements in the list that come before the specified cursor."
-    // before: String): BakerTransactionRelationConnection
-}
-
-#[derive(Union)]
-enum BakerState<'a> {
-    ActiveBakerState(ActiveBakerState<'a>),
-    RemovedBakerState(RemovedBakerState),
-}
-
-#[derive(SimpleObject)]
-struct ActiveBakerState<'a> {
-    // /// The status of the bakers node. Will be null if no status for the node
-    // /// exists.
-    // node_status:      NodeStatus,
-    staked_amount:    Amount,
-    restake_earnings: bool,
-    pool:             BakerPool<'a>,
-    // This will not be used starting from P7
-    pending_change:   Option<PendingBakerChange>,
-}
-
-#[derive(Union)]
-enum PendingBakerChange {
-    PendingBakerRemoval(PendingBakerRemoval),
-    PendingBakerReduceStake(PendingBakerReduceStake),
-}
-
-#[derive(SimpleObject)]
-struct PendingBakerRemoval {
-    effective_time: DateTime,
-}
-
-#[derive(SimpleObject)]
-struct PendingBakerReduceStake {
-    new_staked_amount: Amount,
-    effective_time:    DateTime,
-}
-
-#[derive(SimpleObject)]
-struct RemovedBakerState {
-    removed_at: DateTime,
-}
-
 #[derive(SimpleObject)]
 struct NodeStatus {
     // TODO: add below fields
@@ -2058,49 +1892,6 @@ struct NodeStatus {
     // averageBytesPerSecondIn: Float!
     // averageBytesPerSecondOut: Float!
     id: types::ID,
-}
-
-#[derive(SimpleObject)]
-struct BakerPool<'a> {
-    // /// Total stake of the baker pool as a percentage of all CCDs in existence.
-    // /// Value may be null for brand new bakers where statistics have not
-    // /// been calculated yet. This should be rare and only a temporary
-    // /// condition.
-    // total_stake_percentage:  Decimal,
-    // lottery_power:           Decimal,
-    // payday_commission_rates: CommissionRates,
-    open_status:      Option<BakerPoolOpenStatus>,
-    commission_rates: CommissionRates,
-    metadata_url:     Option<&'a str>,
-    // /// The total amount staked by delegation to this baker pool.
-    // delegated_stake:         Amount,
-    // /// The maximum amount that may be delegated to the pool, accounting for
-    // /// leverage and stake limits.
-    // delegated_stake_cap:     Amount,
-    // /// The total amount staked in this baker pool. Includes both baker stake
-    // /// and delegated stake.
-    // total_stake:             Amount,
-    // delegator_count:         i32,
-    // /// Ranking of the baker pool by total staked amount. Value may be null for
-    // /// brand new bakers where statistics have not been calculated yet. This
-    // /// should be rare and only a temporary condition.
-    // ranking_by_total_stake:  Ranking,
-    // TODO: apy(period: ApyPeriod!): PoolApy!
-    // TODO: delegators("Returns the first _n_ elements from the list." first: Int "Returns the
-    // elements in the list that come after the specified cursor." after: String "Returns the last
-    // _n_ elements from the list." last: Int "Returns the elements in the list that come before
-    // the specified cursor." before: String): DelegatorsConnection
-    // TODO: poolRewards("Returns the first _n_ elements from the list." first: Int "Returns the
-    // elements in the list that come after the specified cursor." after: String "Returns the last
-    // _n_ elements from the list." last: Int "Returns the elements in the list that come before
-    // the specified cursor." before: String): PaydayPoolRewardConnection
-}
-
-#[derive(SimpleObject)]
-struct CommissionRates {
-    transaction_commission:  Option<Decimal>,
-    finalization_commission: Option<Decimal>,
-    baking_commission:       Option<Decimal>,
 }
 
 #[derive(SimpleObject)]
@@ -2247,29 +2038,6 @@ struct AccountFilterInput {
     is_delegator: bool,
 }
 
-#[derive(InputObject)]
-struct BakerFilterInput {
-    open_status_filter: BakerPoolOpenStatus,
-    include_removed:    bool,
-}
-
-#[derive(Enum, Clone, Copy, PartialEq, Eq, Default)]
-enum BakerSort {
-    #[default]
-    BakerIdAsc,
-    BakerIdDesc,
-    BakerStakedAmountAsc,
-    BakerStakedAmountDesc,
-    TotalStakedAmountAsc,
-    TotalStakedAmountDesc,
-    DelegatorCountAsc,
-    DelegatorCountDesc,
-    BakerApy30DaysDesc,
-    DelegatorApy30DaysDesc,
-    BlockCommissionsAsc,
-    BlockCommissionsDesc,
-}
-
 struct SearchResult {
     _query: String,
 }
@@ -2359,7 +2127,7 @@ impl SearchResult {
         #[graphql(desc = "Returns the last _n_ elements from the list.")] _last: Option<i32>,
         #[graphql(desc = "Returns the elements in the list that come before the specified cursor.")]
         _before: Option<String>,
-    ) -> ApiResult<connection::Connection<String, Baker>> {
+    ) -> ApiResult<connection::Connection<String, baker::Baker>> {
         todo_api!()
     }
